requirements.txt
click==8.1.3
datasets==2.13.0
numpy==1.22.4
scikit-learn==1.2.2
scipy==1.10.1
tensorboardX==2.6.1
torchsummary==1.5.1
tqdm==4.65.0
transformers==4.30.2
accelerate==0.20.3
fuzzywuzzy==0.18.0
memory_profiler==0.61.0
xformers==0.0.20
GPUtil==1.4.0